Reorganize and fix mistral unit tests
Separate mistral unit tests into separate files. Move workflow fixtures into a test pack. Fix a bug in the tests when comparing workflow definition on save.
