+++ contrib/runners/mistral_v2/tests/unit/test_mistral_v2_rerun.py
@@ -31,7 +31,7 @@ tests_config.parse_args()
 
 from mistral_v2 import MistralRunner
 from st2common.bootstrap import actionsregistrar
-import st2common.bootstrap.runnersregistrar as runners_registrar
+from st2common.bootstrap import runnersregistrar
 from st2common.constants import action as action_constants
 from st2common.models.db.liveaction import LiveActionDB
 from st2common.persistence.liveaction import LiveAction
@@ -74,7 +74,8 @@ WB1_META_FILE_PATH = TEST_PACK_PATH + '/actions/' + WB1_META_FILE_NAME
 WB1_META_CONTENT = loader.load_meta_file(WB1_META_FILE_PATH)
 WB1_NAME = WB1_META_CONTENT['pack'] + '.' + WB1_META_CONTENT['name']
 WB1_ENTRY_POINT = TEST_PACK_PATH + '/actions/' + WB1_META_CONTENT['entry_point']
-WB1_SPEC = yaml.safe_load(MistralRunner.get_workflow_definition(WB1_ENTRY_POINT))
+WB1_ENTRY_POINT_X = WB1_ENTRY_POINT.replace(WB1_META_FILE_NAME, 'xformed_' + WB1_META_FILE_NAME)
+WB1_SPEC = yaml.safe_load(MistralRunner.get_workflow_definition(WB1_ENTRY_POINT_X))
 WB1_YAML = yaml.safe_dump(WB1_SPEC, default_flow_style=False)
 WB1 = workbooks.Workbook(None, {'name': WB1_NAME, 'definition': WB1_YAML})
 WB1_MAIN_EXEC = {'id': str(uuid.uuid4()), 'state': 'RUNNING'}
@@ -98,7 +99,8 @@ WF1_META_FILE_PATH = TEST_PACK_PATH + '/actions/' + WF1_META_FILE_NAME
 WF1_META_CONTENT = loader.load_meta_file(WF1_META_FILE_PATH)
 WF1_NAME = WF1_META_CONTENT['pack'] + '.' + WF1_META_CONTENT['name']
 WF1_ENTRY_POINT = TEST_PACK_PATH + '/actions/' + WF1_META_CONTENT['entry_point']
-WF1_SPEC = yaml.safe_load(MistralRunner.get_workflow_definition(WF1_ENTRY_POINT))
+WF1_ENTRY_POINT_X = WF1_ENTRY_POINT.replace(WF1_META_FILE_NAME, 'xformed_' + WF1_META_FILE_NAME)
+WF1_SPEC = yaml.safe_load(MistralRunner.get_workflow_definition(WF1_ENTRY_POINT_X))
 WF1_YAML = yaml.safe_dump(WF1_SPEC, default_flow_style=False)
 WF1 = workflows.Workflow(None, {'name': WF1_NAME, 'definition': WF1_YAML})
 WF1_EXEC = {'id': str(uuid.uuid4()), 'state': 'ERROR', 'workflow_name': WF1_NAME}
@@ -132,22 +134,22 @@ class MistralRunnerTest(DbTestCase):
         cfg.CONF.set_override('retry_exp_msec', 100, group='mistral')
         cfg.CONF.set_override('retry_exp_max_msec', 200, group='mistral')
         cfg.CONF.set_override('retry_stop_max_msec', 200, group='mistral')
+        cfg.CONF.set_override('api_url', 'http://0.0.0.0:9101', group='auth')
 
         # Register runners.
-        runners_registrar.register_runners()
+        runnersregistrar.register_runners()
 
         # Register test pack(s).
-        registrar = actionsregistrar.ActionsRegistrar(
+        actions_registrar = actionsregistrar.ActionsRegistrar(
             use_pack_cache=False,
             fail_on_failure=True
         )
 
         for pack in PACKS:
-            registrar.register_from_pack(pack)
+            actions_registrar.register_from_pack(pack)
 
     def setUp(self):
         super(MistralRunnerTest, self).setUp()
-        cfg.CONF.set_override('api_url', 'http://0.0.0.0:9101', group='auth')
 
         # Mock the local runner run method.
         local_runner_cls = self.get_runner_class('local_runner')
