+++ st2tests/st2tests/fixtures/packs/mistral_tests/actions/workflows/workbook_v2_many_workflows_no_default.yaml
@@ -3,47 +3,47 @@ version: '2.0'
 
 workflows:
 
-  main1:
-    type: direct
-    input:
-      - count
-      - friend
-    tasks:
-      greet:
-        workflow: subflow1 count=<% $.count %> friend=<% $.friend %> 
-        publish:
-          towhom: <% task(greet).result.towhom %>
-
-  main2:
-    type: direct
-    input:
-      - count
-      - friend
-    tasks:
-      greet:
-        workflow: subflow1 count=<% $.count %> friend=<% $.friend %>
-        publish:
-          towhom: <% task(greet).result.towhom %>
-
-  subflow1:
-    type: direct
-    input:
-      - count
-      - friend
-    output:
-      towhom: <% $.towhom %>
-    tasks:
-      say-greeting:
-        action: core.hey
+    main1:
+        type: direct
         input:
-          cmd: <% $.count %>
-        publish:
-          greet: <% task(say-greeting).result.stdout %>
-        on-success:
-          - say-friend
-      say-friend:
-        action: core.friend
+            - count
+            - friend
+        tasks:
+            greet:
+                workflow: subflow1 count=<% $.count %> friend=<% $.friend %> 
+                publish:
+                    towhom: <% task(greet).result.towhom %>
+
+    main2:
+        type: direct
         input:
-          cmd: <% $.friend %>
-        publish:
-          towhom: <% task(say-friend).result.stdout %>
+            - count
+            - friend
+        tasks:
+            greet:
+                workflow: subflow1 count=<% $.count %> friend=<% $.friend %>
+                publish:
+                    towhom: <% task(greet).result.towhom %>
+
+    subflow1:
+        type: direct
+        input:
+            - count
+            - friend
+        output:
+            towhom: <% $.towhom %>
+        tasks:
+            say-greeting:
+                action: core.local
+                input:
+                    cmd: <% $.count %>
+                publish:
+                    greet: <% task(say-greeting).result.stdout %>
+                on-success:
+                    - say-friend
+            say-friend:
+                action: core.local
+                input:
+                    cmd: <% $.friend %>
+                publish:
+                    towhom: <% task(say-friend).result.stdout %>
