+++ st2tests/st2tests/fixtures/packs/mistral_tests/actions/workflows/xformed_workbook_v2_many_workflows.yaml
@@ -23,16 +23,20 @@ workflows:
             towhom: <% $.towhom %>
         tasks:
             say-greeting:
-                action: core.local
+                action: st2.action
                 input:
-                    cmd: <% $.count %>
+                    ref: core.local
+                    parameters:
+                        cmd: <% $.count %>
                 publish:
                     greet: <% task(say-greeting).result.stdout %>
                 on-success:
                     - say-friend
             say-friend:
-                action: core.local
+                action: st2.action
                 input:
-                    cmd: <% $.friend %>
+                    ref: core.local
+                    parameters:
+                        cmd: <% $.friend %>
                 publish:
                     towhom: <% task(say-friend).result.stdout %>
