+++ st2tests/st2tests/fixtures/packs/mistral_tests/actions/workflows/xformed_workflow_v2_many_workflows.yaml
@@ -7,17 +7,21 @@ mistral_tests.workflow_v2_many_workflows.main1:
         - friend
     tasks:
         say-greeting:
-            action: core.local
+            action: st2.action
             input:
-                cmd: <% $.count %>
+                ref: core.local
+                parameters:
+                    cmd: <% $.count %>
             publish:
                 greet: <% task(say-greeting).result.stdout %>
             on-success:
                 - say-friend
         say-friend:
-            action: core.local
+            action: st2.action
             input:
-                cmd: <% $.friend %>
+                ref: core.local
+                parameters:
+                    cmd: <% $.friend %>
             publish:
                 towhom: <% task(say-friend).result.stdout %>
 
@@ -28,16 +32,20 @@ mistral_tests.workflow_v2_many_workflows.main2:
         - friend
     tasks:
         say-greeting:
-            action: core.local
+            action: st2.action
             input:
-                cmd: <% $.count %>
+                ref: core.local
+                parameters:
+                    cmd: <% $.count %>
             publish:
                 greet: <% task(say-greeting).result.stdout %>
             on-success:
                 - say-friend
         say-friend:
-            action: core.local
+            action: st2.action
             input:
-                cmd: <% $.friend %>
+                ref: core.local
+                parameters:
+                    cmd: <% $.friend %>
             publish:
                 towhom: <% task(say-friend).result.stdout %>
